embedding_indexer/searchbox_indexer/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-default-group==1.2.4
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
gpt4all==2.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
llm==0.11
llm-embed-jina==0.1.2
llm-gpt4all==0.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openai==0.28.1
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ulid==2.2.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlite-fts4==1.0.3
sqlite-migrate==0.1a2
sqlite-utils==3.35.1
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
types-psycopg2==2.9.21.14
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
